+++ facts.d/JAVA_HOME.sh
@@ -1,2 +1,0 @@
-#! /bin/bash
-echo "machine_java_home=$JAVA_HOME"